+++ srcs/vector/angle.c
@@ -6,7 +6,7 @@
 /*   By: toliver <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/09/19 15:04:58 by toliver           #+#    #+#             */
-/*   Updated: 2018/10/09 13:57:56 by bede-fre         ###   ########.fr       */
+/*   Updated: 2018/10/19 09:35:02 by lguiller         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 float	ft_degtorad(float angle)
 {
-	return (angle / 180.0 * M_PI);
+	return (angle / 180.0f * (float)M_PI);
 }
 
 float	ft_radtodeg(float angle)
 {
-	return (angle * 180 / M_PI);
+	return (angle * 180.0f / (float)M_PI);
 }
